refactor(structure): professionalize folder structure and fix manifest generation
This commit addresses user feedback to professionalize the project structure and fix a build-time error related to manifest generation.

- The folder structure has been reorganized. All sample pages (`qo_*.tsx`) have been moved from the root into a dedicated `src/pages/samples/` directory.

- The manifest generation logic has been moved into `vite.config.js`. This ensures that the `.generated/pages.manifest.json` file is created before Vite attempts to build the application, resolving the "Failed to resolve import" error.

- All related configuration files (`tailwind.config.js`, `tsconfig.json`) and the `README.md` have been updated to reflect these changes.

**Note on Verification:**
As previously noted, the execution environment has severe limitations that prevent running the Vite build process. This commit could not be fully verified, but it directly addresses the error reported by the user and implements the requested project restructuring.
