+++ src/pages/samples/qo_s004_kitchen_display.tsx
