+++ tailwind.config.js
@@ -3,7 +3,7 @@ module.exports = {
   content: [
     "./index.html",
     "./src/**/*.{js,ts,jsx,tsx}",
-    "./qo_*.{js,ts,jsx,tsx}",
+    "./src/pages/samples/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {},
